Move plan to scraper package
