+++ pkg/scraper/plan/memory/helpers.go
