+++ pkg/scraper/plan/memory/helpers_test.go
