+++ pkg/scraper/plan/testdata/resources.yaml
